Refactoring into new class, IndexUtil; still wip.
